Facebook login
Update signinview, test facebook login,
