+++ Frontend/index.jade
@@ -4,6 +4,7 @@ html(lang="en", ng-app="myApp")
         meta(charset='utf8')
         base(href='/')
         link(rel="stylesheet" href="frameworks/bootstrap/3.3.5/css/bootstrap.min.css")
+        link(rel="stylesheet" href="//netdna.bootstrapcdn.com/font-awesome/4.0.3/css/font-awesome.min.css")
         script(src="frameworks/bootstrap/3.3.5/js/bootstrap.min.js")
         script(src="frameworks/angularjs/1.3.16/angular.min.js")
         script(src="frameworks/angularjs/1.3.16/angular-route.min.js")
