+++ Frontend/partials/signin.jade
@@ -1,6 +1,8 @@
 .jumbotron.text-center
     h1
-        span.fa.fa-lock Oak Bay Rec Centre
-    a.btn.btn-default(href="/login") 
-        span.fa.fa-user Local Login
-
+        span.fa.fa-lock  Oak Bay Recreation
+               
+    p Login or Register with:
+               
+    a.btn.btn-primary(href='/auth/facebook')
+        span.fa.fa-facebook  Facebook
