+++ Server/passport/passport.js
@@ -1,9 +1,12 @@
 var mongoose = require('mongoose');
 var LocalStrategy = require('passport-local').Strategy;
+var FacebookStrategy = require('passport-facebook').Strategy;
+var GoogleStrategy = require('passport-google-oauth').OAuth2Strategy;
 var User = mongoose.model('User');
 var config = require('config');
 
 var google = require('./google');
+var facebook = require('./facebook');
 /**
  * Expose
  */
@@ -19,8 +22,11 @@ module.exports = function (passport, app) {
               done(err, user)
           })
     })
+    
+    
+    
     passport.use(google);
-
+    passport.use(facebook); //?? will this work tho
     // Initialize Passport!  Also use passport.session() middleware, to support
     // persistent login sessions (recommended).
     app.use(passport.initialize());
@@ -37,9 +43,42 @@ module.exports = function (passport, app) {
             res.redirect('/');
         }
     );
+    
+    
+    //routes for facebook authentication
+    app.get('/auth/facebook', passport.authenticate('facebook', { scope : 'email' }));
+
+    // handle the callback after facebook has authenticated the user
+    app.get('/auth/facebook/callback',
+        passport.authenticate('facebook', {
+            successRedirect : '/',
+            failureRedirect : '/login'
+        }));
+
+    
+    
+    
+    
+    
+    
+    //route for logging out
 
     app.get('/logout', function(req, res){
       req.logout();
       res.redirect('/');
     })
 };
+
+
+
+
+// route middleware to make sure a user is logged in
+function isLoggedIn(req, res, next) {
+
+    // if user is authenticated in the session, carry on
+    if (req.isAuthenticated())
+        return next();
+
+    // if they aren't redirect them to the home page
+    res.redirect('/');
+}
