+++ package.json
@@ -20,7 +20,7 @@
     "passport-local": "^1.0.0",
     "serve-favicon": "~2.2.0",
     "socket.io": "^1.3.5",   
-    "ejs" : "~0.8.5", 		
+		
     "passport-facebook" : "~1.0.2",
     "passport-twitter" : "~1.0.2",
